Custom tags: Split source files
