+++ CMakeLists.txt
@@ -854,6 +854,8 @@ add_library(mixxx-lib STATIC EXCLUDE_FROM_ALL
   src/track/keyutils.cpp
   src/library/tags/customtags.cpp
   src/library/tags/tag.cpp
+  src/library/tags/tagfacet.cpp
+  src/library/tags/taglabel.cpp
   src/track/playcounter.cpp
   src/track/replaygain.cpp
   src/track/serato/beatgrid.cpp
