+++ src/test/librarytags_test.cpp
@@ -4,6 +4,7 @@
 #include <QJsonObject>
 
 #include "library/tags/tag.h"
+#include "library/tags/tagfacet.h"
 
 namespace mixxx {
